requirements.txt
gitdb==4.0.11
GitPython==3.1.43
Markdown==3.7
python-frontmatter==1.1.0
PyYAML==6.0.2
smmap==5.0.1
